third_party/WebKit/LayoutTests/fast/dom/custom/prerender-insert-after-stop-expected.txt

CONSOLE WARNING: line 4: document.registerElement is deprecated and will be removed in M73, around March 2019. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 for more details.
Test that Custom Element should be defined while navigating between frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS customBody is defined.
PASS customBody.toLocaleString() is "[object HTMLBodyElement]"
PASS successfullyParsed is true

TEST COMPLETE
